fix mongodb circular dependencies and slate build
