+++ packages/db-mongodb/src/queries/parseParams.ts
@@ -4,7 +4,6 @@ import type { Field, Operator, Payload, Where } from 'payload'
 import { deepMergeWithCombinedArrays } from 'payload'
 import { validOperators } from 'payload/shared'
 
-import { buildAndOrConditions } from './buildAndOrConditions.js'
 import { buildSearchParam } from './buildSearchParams.js'
 
 export async function parseParams({
@@ -84,4 +83,42 @@ export async function parseParams({
   }
 
   return result
+}
+
+export async function buildAndOrConditions({
+  collectionSlug,
+  fields,
+  globalSlug,
+  locale,
+  payload,
+  where,
+}: {
+  collectionSlug?: string
+  fields: Field[]
+  globalSlug?: string
+  locale?: string
+  payload: Payload
+  where: Where[]
+}): Promise<Record<string, unknown>[]> {
+  const completedConditions = []
+  // Loop over all AND / OR operations and add them to the AND / OR query param
+  // Operations should come through as an array
+
+  for (const condition of where) {
+    // If the operation is properly formatted as an object
+    if (typeof condition === 'object') {
+      const result = await parseParams({
+        collectionSlug,
+        fields,
+        globalSlug,
+        locale,
+        payload,
+        where: condition,
+      })
+      if (Object.keys(result).length > 0) {
+        completedConditions.push(result)
+      }
+    }
+  }
+  return completedConditions
 }
